nove funkcionalnosti

+++ Database-access/DTO/PlayerFightDto.cs
@@ -1,7 +1,7 @@
 public class PlayerFightCreateDto
 {
     public string Winner { get; set; }
-    public int Experience { get; set; }
+    public double Experience { get; set; }
     public int Honor { get; set; }
     public string Player1Name { get; set; }
     public string Player2Name { get; set; }
@@ -10,6 +10,6 @@ public class PlayerFightUpdateDto
 {
     public int PlayerFightId { get; set; }
     public string Winner { get; set; }
-    public int Experience { get; set; }
+    public double Experience { get; set; }
     public int Honor { get; set; }
 }
